Rename device_resource -> device_child
