+++ src/sgl/device/device_child.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "device_resource.h"
+#include "device_child.h"
 
 namespace sgl {
