+++ src/sgl/device/hot_reload.h
@@ -5,7 +5,7 @@
 #include "sgl/device/fwd.h"
 #include "sgl/device/types.h"
 #include "sgl/device/reflection.h"
-#include "sgl/device/device_resource.h"
+#include "sgl/device/device_child.h"
 
 #include "sgl/core/fwd.h"
 #include "sgl/core/object.h"
