+++ src/sgl/device/query.h
@@ -4,7 +4,7 @@
 
 #include "sgl/device/fwd.h"
 #include "sgl/device/types.h"
-#include "sgl/device/device_resource.h"
+#include "sgl/device/device_child.h"
 
 #include "sgl/core/macros.h"
 #include "sgl/core/object.h"
