+++ src/sgl/device/resource.h
@@ -4,7 +4,7 @@
 
 #include "sgl/device/fwd.h"
 #include "sgl/device/types.h"
-#include "sgl/device/device_resource.h"
+#include "sgl/device/device_child.h"
 #include "sgl/device/formats.h"
 #include "sgl/device/native_handle.h"
