+++ src/slangpy_ext/CMakeLists.txt
@@ -20,7 +20,7 @@ nanobind_add_module(slangpy_ext NB_STATIC LTO
     device/command.cpp
     device/coopvec.cpp
     device/cursor_utils.h
-    device/device_resource.cpp
+    device/device_child.cpp
     device/device.cpp
     device/fence.cpp
     device/formats.cpp
